# services/shared/openapi/auth.yaml
# OpenAPI спецификация для Users Service (Auth)

openapi: 3.0.3
info:
  title: Users Service API
  description: API для управления пользователями и аутентификации
  version: 1.0.0

servers:
  - url: http://localhost:8001
    description: Local development

tags:
  - name: auth
    description: Аутентификация и авторизация
  - name: users
    description: Управление пользователями

paths:
  /auth/register:
    post:
      tags:
        - auth
      summary: Регистрация пользователя
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '201':
          description: Пользователь успешно зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Пользователь уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_code: "USER_ALREADY_EXISTS"
                message: "Пользователь с таким email уже существует"

  /auth/login:
    post:
      tags:
        - auth
      summary: Аутентификация пользователя
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Неверные учетные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_code: "INVALID_CREDENTIALS"
                message: "Неверный email или пароль"

  /auth/refresh:
    post:
      tags:
        - auth
      summary: Обновление access token
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshToken'
      responses:
        '200':
          description: Токены успешно обновлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Недействительный refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_code: "REFRESH_TOKEN_INVALID"
                message: "Недействительный или просроченный refresh token"

  /users/me:
    get:
      tags:
        - users
      summary: Получение профиля текущего пользователя
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    UserRegister:
      type: object
      required:
        - email
        - password
        - role
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          example: "securePassword123"
        role:
          type: string
          enum:
            - USER
            - SELLER
            - ADMIN
          example: "USER"
        first_name:
          type: string
          maxLength: 100
          example: "Иван"
        last_name:
          type: string
          maxLength: 100
          example: "Иванов"

    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          example: "securePassword123"

    RefreshToken:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."

    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          default: "Bearer"
        expires_in:
          type: integer
          description: Время жизни access token в секундах
          example: 1800

    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          example: "user@example.com"
        role:
          type: string
          enum:
            - USER
            - SELLER
            - ADMIN
          example: "USER"
        first_name:
          type: string
          example: "Иван"
        last_name:
          type: string
          example: "Иванов"
        created_at:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00Z"

    ErrorResponse:
      type: object
      required:
        - error_code
        - message
      properties:
        error_code:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "Ошибка валидации входных данных"
        details:
          type: object
          additionalProperties: true

  responses:
    BadRequest:
      description: Ошибка валидации
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Неавторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'